Fixed the messaging bug and created test files
This commit fixes a bug in the core_utils.py where the function fcm.appendMessageToWhatsappNumber() was wrongly nested under the else clause. This caused messages not to be appended for new users. It has been taken out of the else clause to work for both new and existing users.

Two new testing files requisitionBoilerPlate.py and twilioTests.py have been created. They include boiler plate function for Twilio and DialogFlow, and functions to send test requests to each API.

In api.py, the lines to convert and emit user messages have been commented out for testing purposes, but will be restored in production.