Refactor ngrok integration and DialogueFlow fulfillment setting
The ngrok functionality, previously contained in a single file (ngrokRunner.py), has been refactored to provide clearer separation of responsibilities. The new file ngrokGetter.py is responsible for retrieving the ngrok URL, while the launching of the ngrok process has been moved to a different part of the system to avoid unnecessary interconnection.

The file dialogflowFulfillment.py was renamed to _fulfillmentSetter.py and the implementation was changed to return the response of the fulfillment update for more flexibility and better debugging. Also, a new file dialogflowAutomatedFulfillment.py was created to automate the process of setting the fulfillment.

This refactoring provides a more modular, maintainable codebase where the responsibilities of each file (and the system as a whole) are better defined, making it easier to understand and modify in the future.