"Add checkbox to toggle pizza options in chatbot GUI"
In this commit, a checkbox was added to the pizza selection area of the chatbot GUI. This checkbox allows the user to toggle the pizza options on and off. On checking the box, the dropdowns for the pizza options reset to their default values. On unchecking the box, the dropdowns are cleared. This feature provides a dynamic user experience where the user has control over the pizza options dropdown list. The overall layout of the dropdowns was also adjusted with grid positioning for better visual harmony.