Update UI elements for improved Pizza option handling
Modified the GUI to handle the 'Pizza Choose' option separately in the main window. This change was made because the method to update the text field for the option handles differently than other options.

In `_dropdown_component.py`, the DropdownComponent was changed to instance variable for further use.

In `_option_frame.py`, the 'Pizza Choose' option was handled separately, with options for both 'First Pizza' and 'Second Pizza'. This improves the usability of the 'Pizza Choose' option by allowing for separate selection of each pizza type. The chosen options are now directly set into `pizza_dropdown_values`.

In `main_ui.py`, added an extra condition to handle if the chosen option from the dropdown menu is "2- Pizza Choose". If this is the case, the new method `update_pizza_text()` from `_option_frame.py` is called. For all other cases, the previous behavior is maintained.