Refactor pizza selection logic and enhance code readability
The pizza selection logic in _option_frame.py was confusing and unnecessarily complicated. It has been refactored into a separate private function, __pizzaChooseLogic, to improve readability and modularity. This also includes assigning the initial pizza values, which were hardcoded, to the first four pizza flavors in pizzaFlavors list. The mapping of widget names to indices was used using a dictionary to eliminate the if-else branching, further improving the code readability while simultaneously streamlining index assignment. A minor typo in a variable name, iSecondPizzaHomogeneous, was corrected to isSecondPizzaHomogeneous to maintain the naming consistency.