Update README.md with steps for running the server and setting webhooks
Minor corrections were made to README.md to improve grammar and consistency, but its primary update is the addition of in-depth instructions for running the server and setting up webhooks. This enrichment of the originally sparse instructions was necessary to promote user accessibility and clarity, thus allowing a wider range of developers to utilize the program effectively. Comprehensive steps, supplemented by visuals, have been provided for running the server, exposing the url using ngrok, and manually setting up dialogflow, twilio, and instagram webhooks.