Reformat README & add Dialogflow configuration guide
The commit primarily serves two purposes. Firstly, it corrects the Helper Component heading from a primary header to a secondary header, which is a better representation of the document structure. Secondly, it introduces a detailed guide on setting up the Dialogflow configuration for a ChatBot. This guide, divided into 5 comprehensive steps, will streamline the process for new users who are not very familiar with the working of Dialogflow or the structure of the project. Lastly, a minor change relates to inserting a line break for the 'webhooks' section to improve readability.