Refactor deployment workflow for efficiency
The sequence of the GitHub action steps has been rearranged for increased readability and efficiency. Actions for Google Cloud authentication and Docker image building and pushing have been moved above the environment variable settings. The use of an array for environment variables has been implemented for better environment variable management. The UPDATE_ENV_VARS_ARG is then used in the deployment to unify the application of environment variables which significantly streamlines the workflow.