Improve readability and modularity of chatbot API and test cases
This commit makes several changes to improve the readability and modularity of the chatbot API. Importantly, commented out code related to gevent monkey patching has been uncommented to support concurrency. Test-related routes and functions have been moved closer together in 'api.py' for better organization. Environment variable across files has been updated for consistency (from "DIALOGFLOW_PROJECT_ID" to "SDK_PROJECT_ID"). The session IDs are now generated with the 'uuid' library, providing unique session IDs for Dialogflow sessions. Unused code has been removed, and print statements are added for better debugging. In 'core_utils.py', the main function is simplified, and a test message is updated. Finally, in the orderProcessor and fulfillment components, minor changes are made that optimize and clarify the code. These changes improve the overall quality of the codebase, making it easier to understand and work with. Also, these changes ensure that the system works stably with multi-threading and allows for easier debugging and testing of the code.