Update import statement and clean up unused code
Modified how datetime is imported in the firebase_conversation.py file to enhance clarity of usage. This change simplifies the timestamping of messages. In addition, some commented-out code was removed for code cleanliness and overall readability, this code described dummy conversations creation and is unnecessary for the operation of the framework.

Moreover, extraneous return of the conversation variable has been eliminated from the updateFirebaseWithUserMessage function in core_utils.py to enhance simplicity, as this data was not used.

Fulfillment url has also been changed to point to cloud run instead of ngrok in the fulfillmentModule. This configuration changes have been made to support a shift in the development environment.

Also, changes have been made in api.py to refactor the entire process of handling incoming messages and bot responses, which enhance code clarity and maintainability.