Refactor Firebase append messages function
Removed commented code in the appendMultipleMessagesToFirebase function which resulted in making the code cleaner. Also, separated the headers and payload within the makeHttpCallToAppendMultipleMessagesToFirebaseServerlessFunction function to improve readability and maintainability of the code. The payload is also now explicitly defined as a JSON object. This change enhances the code performance by ensuring the data sent to the Firebase Serverless function is in the correct format.