Add caching functionality for the Firebase data
Implemented functionality to cache Firebase data in a JSON file to reduce the number of unnecessary read operations and to speed up subsequent data accesses. Modified .gitignore to ignore the new cache files. Additionally, a helper function is added in a newly created utils file, date_utils.py, that converts a time delta to a string, which is used to print the duration since the last cache update. Furthermore, removed a firebase function call in api.py temporarily in order to debug the implemented changes. Made changes to speisekarte.json to test the new caching functionality.