"Added speisekarte routes and updated API configuration file
In this commit, we introduced the 'speisekarte' (menu) routes that provide a set of endpoints for creating, retrieving, updating and deleting menus. This expansion of routes is added to support additional functionality in a restaurant ordering application. The api.py file has been updated to include the new speisekarte routes.

We also added the FirebaseSpeisekarte class to the api_config.py file to handle the database operations related to the 'speisekarte'. This allows our application to interact with the speisekarte in the database.

The change is important to extend the capabilities of our restaurant application, enabling it to handle operations related to menus."