"Added order routes and updated API configuration file
In this commit, we created the 'order' routes that provide a set of endpoints for creating, retrieving, updating and deleting orders. Additionally, we created `firebase_order.py` file where all order related database operations are handled. This expansion of routes and database operations are added to support overall functionality in the application. The `api.py` and `api_config.py` files have been updated to include the new order routes and `FirebaseOrder` class respectively.

The changes facilitate smoother interaction between the application and the database for order related operations and expand the functionality of our ordering system."