Remove unnecessary commented code and add firebase push ID generator
This commit removes some commented out code in the `firebaseCache/cache_utils.py` file that appeared to be related to printing timestamp information for a cache, which was presumably used for debugging and is no longer necessary.

In the `utils/firebase_utils.py` file, a `generate_firebase_push_id` function was added to create unique identifiers for data pushed to Firebase. This uses a combination of the current time and a random number to ensure uniqueness.

Also, a new file `firebaseCache/conversation_cache.py` was added, this file includes a functionality to generate an object that can be used to run operations pertaining to conversations in the local cache, and can be used as a blueprint for conversation related operations.