Refactor order processing and user management
Refactoring done to improve the readability and functionality of the code. Changes made to file `core_intent_processing.py` includes formatting of order data for better readability and processing. In `api.py`, user registration test from `signupBot` was removed and exceptions handling was adjusted for better error tracking and debugging. Furthermore, extracted the user message from the metadata in `core_utils.py` to improve the user experience. In `whatsapp_handle_new_user.py` and `whatsapp_user_manager.py` refactored the way new users are handled and existing users are checked to ensure seamless integration of users in the application. Additionally, refactored the naming conventions in the user manager to fit better with the rest of the application.