Update logger initialization and enhance exception handling
The logger initialization in app.py is updated with format and datefmt. Additionally, the logging.error method is replaced with logging.exception which provides a stack trace in case of an exception. The return 'application/json' text strings are moved into a variable to avoid hard coding the same string multiple times. Commented out code is removed to enhance readability and the function '_get_bot_response_from_user_session' is run in executor for better performance. Also, every exception is now logged before returning a response to better understand issues during runtime. Lastly, a minor change was made in cache_table.json file to update the cache timestamp.