Refactor to streamline creation of SignupBot sessions
The refactoring changes consolidate the creation and initialization process of SignupBot instances in the core_utils.py file, which will improve code maintainability and allow simpler updates to these processes in the future. Additionally, the handleNewWhatsappUser method in the signupBot/whatsapp_handle_new_user.py file was modified to simplify the retrieval of phoneNumber and userMessage, thereby enhancing efficiency. Now this method takes metaData and a SignupBot instance as arguments, which facilitates better session management.